moved mocks to tests/mocks
